updated speed variable for hatch pitch

+++ src/main/java/frc/robot/commands/PitchHatch.java
@@ -5,14 +5,14 @@ import edu.wpi.first.wpilibj.command.Command;
 import frc.robot.Robot;
 
 public class PitchHatch extends Command {
-  public PitchHatch(boolean isForwards, float speed) {
+  public PitchHatch(boolean isForwards, double speed) {
     this.isForwards = isForwards;
     this.speed = speed;
     requires(Robot.hatchPanelSubsystem);
   }
 
   private boolean isForwards;
-  private float speed;
+  private double speed;
 
   @Override
   protected void initialize() {
